Исправление замечаи от "iBerbasov"
