--- src/main/java/school/faang/Book.java
@@ -1,33 +1,0 @@
-package school.faang;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.Objects;
-
-@Getter
-@Builder
-public class Book {
-    private String title;
-    private String author;
-    private int year;
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Book book = (Book) o;
-        return year == book.year
-                && Objects.equals(title, book.title)
-                && Objects.equals(author, book.author);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(title, author, year);
-    }
-}
