+++ src/main/java/school/faang/westeroslibrary/Book.java
@@ -1,0 +1,12 @@
+package school.faang.westeroslibrary;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class Book {
+    private String title;
+    private String author;
+    private int year;
+}
